pyproject.toml
[project]
name = "shopping-agent"
version = "0.1.0"
description = "Shopping agent to automate your shopping!"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arize-phoenix-otel>=0.10.3",
    "gradio>=5.33.1",
    "llama-index>=0.12.41",
    "llama-index-core>=0.12.41",
    "llama-index-llms-gemini>=0.5.0",
    "llama-index-tools-duckduckgo>=0.3.0",
    "markdownify>=1.1.0",
    "python-dotenv>=1.1.0",
]
